add POST endpoint to create a calendar
