+++ packages/cal-api/src/helpers/recurringEntriesHelpers.ts
@@ -1,13 +1,20 @@
-import { CalendarEntry } from "../models/calendarEntry";
-import {type EntryExceptionType, RecurringEntryType} from "../types"
-import { EntryException } from "../models/entryException";
-import { addMillisecondsToDate, dateMinusMinutes, datePlusMinutes, getMillisecondsBetween, getTimeFromDate, setTimeForDate } from "./dateHelpers";
 import { RRule, RRuleSet, rrulestr } from "rrule";
-import { FREQUENCY_MAPPING } from "../controllers/calendar";
+import { FREQUENCY_MAPPING } from "../controllers/calendarEntries";
+import { CalendarEntry } from "../models/calendarEntry";
+import { EntryException } from "../models/entryException";
+import { RecurringEntryType, type EntryExceptionType } from "../types";
+import {
+  addMillisecondsToDate,
+  dateMinusMinutes,
+  datePlusMinutes,
+  getMillisecondsBetween,
+  getTimeFromDate,
+  setTimeForDate,
+} from "./dateHelpers";
 
 export const expandModifiedEntryException = async (
   entryException: EntryExceptionType,
-  parentCalendarEntry: RecurringEntryType,
+  parentCalendarEntry: RecurringEntryType
 ) => {
   return {
     _id: parentCalendarEntry._id,
@@ -40,7 +47,7 @@ export const getExpandedEntryExceptions = async (calendarEntry, start, end) => {
   });
 
   return await Promise.all(promises);
-}
+};
 
 export const addDeletedDatesToRuleSet = async (ruleSet, calendarEntryId) => {
   const deletedExceptions = await EntryException.find({
@@ -51,9 +58,12 @@ export const addDeletedDatesToRuleSet = async (ruleSet, calendarEntryId) => {
   deletedExceptions.forEach((exception) => {
     ruleSet.exdate(exception.startTimeUtc);
   });
-}
+};
 
-export const findMatchingModifiedExceptions = async (start, parentCalendarEntry) => {
+export const findMatchingModifiedExceptions = async (
+  start,
+  parentCalendarEntry
+) => {
   const startDate = new Date(start as string);
   const oneMinBefore = dateMinusMinutes(startDate, 1);
   const oneMinAfter = datePlusMinutes(startDate, 1);
@@ -67,7 +77,13 @@ export const findMatchingModifiedExceptions = async (start, parentCalendarEntry)
     .lt(oneMinAfter);
 };
 
-export const getExpandedRecurringEntries = (ruleSet, calendarEntry, start, end, duration) => {
+export const getExpandedRecurringEntries = (
+  ruleSet,
+  calendarEntry,
+  start,
+  end,
+  duration
+) => {
   const recurrences = ruleSet.between(new Date(start), new Date(end));
 
   return recurrences.map((date) => {
@@ -86,17 +102,17 @@ export const getExpandedRecurringEntries = (ruleSet, calendarEntry, start, end, 
       recurrenceEndsUtc: calendarEntry.recurrenceEndsUtc,
     };
   });
-}
+};
 
 export const getRecurringEntriesWithinRange = async (start, end) => {
   const recurringEntriesBeginsWithin: RecurringEntryType[] =
-  await CalendarEntry.find()
-    .where("recurring")
-    .equals(true)
-    .where("startTimeUtc")
-    .gte(start)
-    .where("startTimeUtc")
-    .lt(end);
+    await CalendarEntry.find()
+      .where("recurring")
+      .equals(true)
+      .where("startTimeUtc")
+      .gte(start)
+      .where("startTimeUtc")
+      .lt(end);
   const recurringEntriesBeginsBefore: RecurringEntryType[] =
     await CalendarEntry.find()
       .where("recurring")
@@ -116,27 +132,37 @@ export const getRecurringEntriesWithinRange = async (start, end) => {
     const expandedEntries = await expandRecurringEntry(
       recurringEntry,
       start,
-      end,
+      end
     );
 
     allRecurrences.push(...expandedEntries);
   }
-  return allRecurrences
-}
+  return allRecurrences;
+};
 
 export const expandRecurringEntry = async (calendarEntry, start, end) => {
   const duration = getMillisecondsBetween(
     calendarEntry.startTimeUtc,
-    calendarEntry.endTimeUtc,
+    calendarEntry.endTimeUtc
   );
   const rule = rrulestr(calendarEntry.recurrencePattern);
   const rruleSet = new RRuleSet();
   rruleSet.rrule(rule);
 
-  await addDeletedDatesToRuleSet(rruleSet, calendarEntry._id)
+  await addDeletedDatesToRuleSet(rruleSet, calendarEntry._id);
 
-  const expandedRecurringEntries = getExpandedRecurringEntries(rruleSet, calendarEntry, start, end, duration)
-  const expandedExceptionEntries = await getExpandedEntryExceptions(calendarEntry, start, end)
+  const expandedRecurringEntries = getExpandedRecurringEntries(
+    rruleSet,
+    calendarEntry,
+    start,
+    end,
+    duration
+  );
+  const expandedExceptionEntries = await getExpandedEntryExceptions(
+    calendarEntry,
+    start,
+    end
+  );
 
   return expandedRecurringEntries.concat(expandedExceptionEntries);
 };
@@ -144,7 +170,7 @@ export const expandRecurringEntry = async (calendarEntry, start, end) => {
 export const handleRecurringEntryDeletion = async (entryToDelete, start) => {
   const existingModifiedExceptions = await findMatchingModifiedExceptions(
     start,
-    entryToDelete,
+    entryToDelete
   );
   if (existingModifiedExceptions.length > 0) {
     existingModifiedExceptions[0].remove();
@@ -156,7 +182,7 @@ export const handleRecurringEntryDeletion = async (entryToDelete, start) => {
       startTimeUtc: start,
     });
   }
-}
+};
 
 export const updateExceptionDetails = (exception, data) => {
   exception.title = data.title;
@@ -166,8 +192,8 @@ export const updateExceptionDetails = (exception, data) => {
   exception.startTimeUtc = data.startTimeUtc;
   exception.endTimeUtc = data.endTimeUtc;
   exception.save();
-  return exception  
-}
+  return exception;
+};
 
 export const updateRecurrenceRule = async (entry) => {
   const rule = new RRule({
@@ -177,7 +203,7 @@ export const updateRecurrenceRule = async (entry) => {
   });
   entry.recurrencePattern = rule.toString();
   await entry.save();
-}
+};
 
 export const createEntryException = async (entry, data) => {
   await EntryException.create({
@@ -197,65 +223,92 @@ export const createEntryException = async (entry, data) => {
     allDay: data.allDay,
     endTimeUtc: data.endTimeUtc,
   });
-}
+};
 
 export const updateRelevantFieldsOnSeries = async (entry, data) => {
-  const {
-    updatedStartUtc,
-    updatedEndUtc
-  } = calculateNewStartAndEnd(entry, data.startTimeUtc, data.endTimeUtc)
+  const { updatedStartUtc, updatedEndUtc } = calculateNewStartAndEnd(
+    entry,
+    data.startTimeUtc,
+    data.endTimeUtc
+  );
 
   // The following are the only fields that are allowed to be edited on a recurring series
-  entry.startTimeUtc = updatedStartUtc
-  entry.endTimeUtc = updatedEndUtc
-  entry.title = data.title
-  entry.description = data.description
-  entry.recurrenceEndsUtc = data.recurrenceEndsUtc
-  entry.frequency = data.frequency
-  entry.allDay = data.allDay
-  await entry.save()
-  await updateRecurrenceRule(entry)
-  return entry
-}
+  entry.startTimeUtc = updatedStartUtc;
+  entry.endTimeUtc = updatedEndUtc;
+  entry.title = data.title;
+  entry.description = data.description;
+  entry.recurrenceEndsUtc = data.recurrenceEndsUtc;
+  entry.frequency = data.frequency;
+  entry.allDay = data.allDay;
+  await entry.save();
+  await updateRecurrenceRule(entry);
+  return entry;
+};
 
 const calculateNewStartAndEnd = (originalEntry, newStart, newEnd) => {
-  const {hours: newStartHours, minutes: newStartMinutes} = getTimeFromDate(newStart)
-  const {hours: newEndHours, minutes: newEndMinutes} = getTimeFromDate(newEnd)
+  const { hours: newStartHours, minutes: newStartMinutes } =
+    getTimeFromDate(newStart);
+  const { hours: newEndHours, minutes: newEndMinutes } =
+    getTimeFromDate(newEnd);
 
-  const updatedStartUtc = setTimeForDate(originalEntry.startTimeUtc, newStartHours, newStartMinutes)
-  const updatedEndUtc = setTimeForDate(originalEntry.endTimeUtc, newEndHours, newEndMinutes)
+  const updatedStartUtc = setTimeForDate(
+    originalEntry.startTimeUtc,
+    newStartHours,
+    newStartMinutes
+  );
+  const updatedEndUtc = setTimeForDate(
+    originalEntry.endTimeUtc,
+    newEndHours,
+    newEndMinutes
+  );
 
   return {
     updatedStartUtc,
-    updatedEndUtc
-  }
-}
+    updatedEndUtc,
+  };
+};
 
 export const updateEntryExceptionsForEntry = async (entry) => {
   const modifiedEntryExceptions = await EntryException.find({
     entryId: entry._id,
     modified: true,
-  })
+  });
 
-  const {hours: newStartHours, minutes: newStartMinutes} = getTimeFromDate(entry.startTimeUtc)
-  const {hours: newEndHours, minutes: newEndMinutes} = getTimeFromDate(entry.endTimeUtc)
+  const { hours: newStartHours, minutes: newStartMinutes } = getTimeFromDate(
+    entry.startTimeUtc
+  );
+  const { hours: newEndHours, minutes: newEndMinutes } = getTimeFromDate(
+    entry.endTimeUtc
+  );
 
   for (const exception of modifiedEntryExceptions) {
-    exception.title = entry.title
-    exception.description = entry.description
-    exception.allDay = entry.allDay
-    exception.startTimeUtc = setTimeForDate(exception.startTimeUtc, newStartHours, newStartMinutes)
-    exception.endTimeUtc = setTimeForDate(exception.endTimeUtc, newEndHours, newEndMinutes)
-    await exception.save()
+    exception.title = entry.title;
+    exception.description = entry.description;
+    exception.allDay = entry.allDay;
+    exception.startTimeUtc = setTimeForDate(
+      exception.startTimeUtc,
+      newStartHours,
+      newStartMinutes
+    );
+    exception.endTimeUtc = setTimeForDate(
+      exception.endTimeUtc,
+      newEndHours,
+      newEndMinutes
+    );
+    await exception.save();
   }
 
   const deletedEntryExceptions = await EntryException.find({
     entryId: entry._id,
     deleted: true,
-  })
+  });
 
   for (const exception of deletedEntryExceptions) {
-    exception.startTimeUtc = setTimeForDate(exception.startTimeUtc, newStartHours, newStartMinutes)
-    await exception.save()
+    exception.startTimeUtc = setTimeForDate(
+      exception.startTimeUtc,
+      newStartHours,
+      newStartMinutes
+    );
+    await exception.save();
   }
-}
+};
