+++ packages/cal-api/src/models/calendar.ts
@@ -1,0 +1,29 @@
+import mongoose from "mongoose";
+
+const schema = new mongoose.Schema(
+  {
+    eventId: {
+      type: String,
+      required: true,
+    },
+    creatorId: {
+      type: String,
+      required: true,
+    },
+    title: {
+      type: String,
+      required: true,
+    },
+    description: {
+      type: String,
+      required: false,
+    },
+  },
+  {
+    timestamps: true,
+  }
+);
+
+const Calendar = mongoose.model("Calendar", schema, "calendars");
+
+export { Calendar };
